face_ethnicity_recognition/requirements.txt
tensorflow==2.13.0
opencv-python==4.8.0.74
dlib
face-recognition
mtcnn
imutils
streamlit==1.26.0
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
numpy==1.24.3
pillow==10.0.0
scikit-learn==1.3.0
